Import DIA-Umpire files

Import DIA-Umpire files

DIAUmpiretoMSstatsFormat(
  raw.frag,
  raw.pep,
  raw.pro,
  annotation,
  useSelectedFrag = TRUE,
  useSelectedPep = TRUE,
  removeFewMeasurements = TRUE,
  removeProtein_with1Feature = FALSE,
  summaryforMultipleRows = max,
  use_log_file = TRUE,
  append = FALSE,
  verbose = TRUE,
  log_file_path = NULL,
  ...
)

Arguments

raw.frag
name of FragSummary_date.xls data, which includes feature-level data.

raw.pep
name of PeptideSummary_date.xls data, which includes selected fragments information.

raw.pro
name of ProteinSummary_date.xls data, which includes selected peptides information.

annotation
name of annotation data which includes Condition, BioReplicate, Run information.

useSelectedFrag
TRUE will use the selected fragment for each peptide. 'Selected_fragments' column is required.

useSelectedPep
TRUE will use the selected peptide for each protein. 'Selected_peptides' column is required.

removeFewMeasurements
TRUE (default) will remove the features that have 1 or 2 measurements across runs.

removeProtein_with1Feature
TRUE will remove the proteins which have only 1 feature, which is the combination of peptide, precursor charge, fragment and charge. FALSE is default.

summaryforMultipleRows
max(default) or sum - when there are multiple measurements for certain feature and certain run, use highest or sum of multiple intensities.

use_log_file
logical. If TRUE, information about data processing will be saved to a file.

append
logical. If TRUE, information about data processing will be added to an existing log file.

verbose
logical. If TRUE, information about data processing wil be printed to the console.

log_file_path
character. Path to a file to which information about data processing will be saved. If not provided, such a file will be created automatically. If `append = TRUE`, has to be a valid path to a file.

...
additional parameters to `data.table::fread`.

Value

data.frame in the MSstats required format.

Examples

diau_frag = system.file("tinytest/raw_data/DIAUmpire/dia_frag.csv",
                             package = "MSstatsConvert")
diau_pept = system.file("tinytest/raw_data/DIAUmpire/dia_pept.csv",
                             package = "MSstatsConvert")
diau_prot = system.file("tinytest/raw_data/DIAUmpire/dia_prot.csv",
                             package = "MSstatsConvert")
annot = system.file("tinytest/annotations/annot_diau.csv",
                    package = "MSstatsdev")
diau_frag = data.table::fread(diau_frag)
diau_pept = data.table::fread(diau_pept)
diau_prot = data.table::fread(diau_prot)
annot = data.table::fread(annot)

diau_imported = DIAUmpiretoMSstatsFormat(diau_frag, diau_pept, diau_prot,
                                         annot, use_log_file = FALSE)

#> INFO  [2021-05-15 00:30:02] ** Raw data from DIAUmpire imported successfully.
#> INFO  [2021-05-15 00:30:02] ** Using selected fragments and peptides.
#> INFO  [2021-05-15 00:30:02] ** Extracted the data from selected fragments and/or peptides.
#> INFO  [2021-05-15 00:30:02] ** Raw data from DIAUmpire cleaned successfully.
#> INFO  [2021-05-15 00:30:02] ** Using provided annotation.
#> INFO  [2021-05-15 00:30:02] ** Run labels were standardized to remove symbols such as '.' or '%'.
#> INFO  [2021-05-15 00:30:02] ** The following options are used:
#>   - Features will be defined by the columns: PeptideSequence, FragmentIon
#>   - Shared peptides will be removed.
#>   - Proteins with single feature will not be removed.
#>   - Features with less than 3 measurements across runs will be removed.
#> INFO  [2021-05-15 00:30:02] ** Features with all missing measurements across runs are removed.
#> INFO  [2021-05-15 00:30:02] ** Shared peptides are removed.

#> Warning: brak argumentów w max; zwracanie wartości -Inf

#> Error in `[.data.table`(input, , list(Intensity = aggregator(Intensity,     na.rm = TRUE)), by = feature_columns): Column 1 of result for group 16 is type 'double' but expecting type 'integer'. Column types must be consistent for each group.

head(diau_imported)

#> Error in head(diau_imported): nie znaleziono obiektu 'diau_imported'