Create output of group comparison based on results for individual proteins

Create output of group comparison based on results for individual proteins

MSstatsGroupComparisonOutput(input, summarization_output, log_base = 2)

Arguments

input
output of MSstatsGroupComparison function

summarization_output
output of dataProcess function

log_base
base of the logarithm used in fold-change calculation

Value

list, same as the output of `groupComparison`

Examples

QuantData <- dataProcess(SRMRawData, use_log_file = FALSE)

#> INFO  [2021-05-14 22:19:19] ** Features with one or two measurements across runs are removed.
#> INFO  [2021-05-14 22:19:19] ** Fractionation handled.
#> INFO  [2021-05-14 22:19:19] ** Updated quantification data to make balanced design. Missing values are marked by NA
#> INFO  [2021-05-14 22:19:19] ** Log2 intensities under cutoff = 3.776  were considered as censored missing values.
#> INFO  [2021-05-14 22:19:19] ** Log2 intensities = NA were considered as censored missing values.
#> INFO  [2021-05-14 22:19:19] ** Use all features that the dataset originally has.
#> INFO  [2021-05-14 22:19:19] 
#>  # proteins: 2
#>  # peptides per protein: 2-2
#>  # features per peptide: 3-3
#> INFO  [2021-05-14 22:19:19] 
#>                     1 2 3 4 5 6 7 8 9 10
#>              # runs 3 3 3 3 3 3 3 3 3  3
#>     # bioreplicates 3 3 3 3 3 3 3 3 3  3
#>  # tech. replicates 1 1 1 1 1 1 1 1 1  1
#> INFO  [2021-05-14 22:19:19]  == Start the summarization per subplot...
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-14 22:19:19]  == Summarization is done.

group_comparison_input = MSstatsPrepareForGroupComparison(QuantData)
levels(QuantData$ProteinLevelData$GROUP)

#>  [1] "1"  "10" "2"  "3"  "4"  "5"  "6"  "7"  "8"  "9" 

comparison <- matrix(c(-1,0,0,0,0,0,1,0,0,0),nrow=1)
row.names(comparison) <- "T7-T1"
groups = levels(QuantData$ProteinLevelData$GROUP)
colnames(comparison) <- groups[order(as.numeric(groups))]
samples_info = getSamplesInfo(QuantData)
repeated = checkRepeatedDesign(QuantData)
group_comparison = MSstatsGroupComparison(group_comparison_input, comparison,
                                          FALSE, repeated, samples_info)

#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%

group_comparison_final = MSstatsGroupComparisonOutput(group_comparison,
                                                      QuantData)
group_comparison_final[["ComparisonResult"]]

#>   Protein Label      log2FC        SE     Tvalue DF       pvalue   adj.pvalue
#> 1    IDHC T7-T1  6.15438359 0.2917031 21.0981085 18 3.819167e-14 7.638334e-14
#> 2    PMG2 T7-T1 -0.09032809 0.1518673 -0.5947831 18 5.593918e-01 5.593918e-01
#>   issue MissingPercentage ImputationPercentage
#> 1    NA                 0                    0
#> 2    NA                 0                    0