Import Skyline files

Import Skyline files

SkylinetoMSstatsFormat(
  input,
  annotation = NULL,
  removeiRT = TRUE,
  filter_with_Qvalue = TRUE,
  qvalue_cutoff = 0.01,
  useUniquePeptide = TRUE,
  removeFewMeasurements = TRUE,
  removeOxidationMpeptides = FALSE,
  removeProtein_with1Feature = FALSE,
  use_log_file = TRUE,
  append = FALSE,
  verbose = TRUE,
  log_file_path = NULL,
  ...
)

Arguments

input
name of MSstats input report from Skyline, which includes feature-level data.

annotation
name of 'annotation.txt' data which includes Condition, BioReplicate, Run. If annotation is already complete in Skyline, use annotation=NULL (default). It will use the annotation information from input.

removeiRT
TRUE (default) will remove the proteins or peptides which are labeld 'iRT' in 'StandardType' column. FALSE will keep them.

filter_with_Qvalue
TRUE(default) will filter out the intensities that have greater than qvalue_cutoff in DetectionQValue column. Those intensities will be replaced with zero and will be considered as censored missing values for imputation purpose.

qvalue_cutoff
Cutoff for DetectionQValue. default is 0.01.

useUniquePeptide
TRUE (default) removes peptides that are assigned for more than one proteins. We assume to use unique peptide for each protein.

removeFewMeasurements
TRUE (default) will remove the features that have 1 or 2 measurements across runs.

removeOxidationMpeptides
TRUE will remove the peptides including 'oxidation (M)' in modification. FALSE is default.

removeProtein_with1Feature
TRUE will remove the proteins which have only 1 feature, which is the combination of peptide, precursor charge, fragment and charge. FALSE is default.

use_log_file
logical. If TRUE, information about data processing will be saved to a file.

append
logical. If TRUE, information about data processing will be added to an existing log file.

verbose
logical. If TRUE, information about data processing wil be printed to the console.

log_file_path
character. Path to a file to which information about data processing will be saved. If not provided, such a file will be created automatically. If `append = TRUE`, has to be a valid path to a file.

...
additional parameters to `data.table::fread`.

Value

data.frame in the MSstats required format.

Examples

skyline_raw = system.file("tinytest/raw_data/Skyline/skyline_input.csv",
                          package = "MSstatsConvert")
skyline_raw = data.table::fread(skyline_raw)
skyline_imported = SkylinetoMSstatsFormat(skyline_raw)

#> INFO  [2021-05-15 00:30:10] ** Raw data from Skyline imported successfully.
#> INFO  [2021-05-15 00:30:10] ** Raw data from Skyline cleaned successfully.
#> INFO  [2021-05-15 00:30:10] ** Using annotation extracted from quantification data.
#> INFO  [2021-05-15 00:30:10] ** Run labels were standardized to remove symbols such as '.' or '%'.
#> INFO  [2021-05-15 00:30:10] ** The following options are used:
#>   - Features will be defined by the columns: PeptideSequence, PrecursorCharge, FragmentIon, ProductCharge
#>   - Shared peptides will be removed.
#>   - Proteins with single feature will not be removed.
#>   - Features with less than 3 measurements across runs will be removed.
#> INFO  [2021-05-15 00:30:10] ** Rows with values of StandardType equal to iRT are removed 
#> INFO  [2021-05-15 00:30:10] ** Intensities with values of Truncated equal to TRUE are replaced with NA
#> WARN  [2021-05-15 00:30:10] ** DetectionQValue not found in input columns.
#> INFO  [2021-05-15 00:30:10] ** Sequences containing DECOY, Decoys are removed.
#> INFO  [2021-05-15 00:30:10] ** Three isotopic preaks per feature and run are summed
#> INFO  [2021-05-15 00:30:10] ** Features with all missing measurements across runs are removed.
#> INFO  [2021-05-15 00:30:10] ** Shared peptides are removed.
#> INFO  [2021-05-15 00:30:10] ** Multiple measurements in a feature and a run are summarized by summaryforMultipleRows: sum
#> INFO  [2021-05-15 00:30:10] ** Features with one or two measurements across runs are removed.
#> INFO  [2021-05-15 00:30:10] ** Run annotation merged with quantification data.
#> INFO  [2021-05-15 00:30:10] ** Features with one or two measurements across runs are removed.
#> INFO  [2021-05-15 00:30:10] ** Fractionation handled.
#> INFO  [2021-05-15 00:30:10] ** Updated quantification data to make balanced design. Missing values are marked by NA
#> INFO  [2021-05-15 00:30:10] ** Finished preprocessing. The dataset is ready to be processed by the dataProcess function.

head(skyline_imported)

#>   ProteinName      PeptideSequence PrecursorCharge FragmentIon ProductCharge
#> 1      P00370 AANAGGVATSGLEMAQNAAR               3          NA            NA
#> 2      P00370 AANAGGVATSGLEMAQNAAR               3          NA            NA
#> 3      P00370 AANAGGVATSGLEMAQNAAR               3          NA            NA
#> 4      P00370 AANAGGVATSGLEMAQNAAR               3          NA            NA
#> 5      P00370 AANAGGVATSGLEMAQNAAR               3          NA            NA
#> 6      P00370 AANAGGVATSGLEMAQNAAR               3          NA            NA
#>   IsotopeLabelType Condition BioReplicate
#> 1            light      Mix1            1
#> 2            light      Mix1            2
#> 3            light      Mix1            3
#> 4            light      Mix2            4
#> 5            light      Mix2            5
#> 6            light      Mix2            6
#>                                              Run Fraction  Intensity
#> 1 121219_S_CCES_01_01_LysC_Try_1to10_Mixt_1_1raw        1 5311459776
#> 2 121219_S_CCES_01_02_LysC_Try_1to10_Mixt_1_2raw        1 4900185344
#> 3 121219_S_CCES_01_03_LysC_Try_1to10_Mixt_1_3raw        1 5323685504
#> 4 121219_S_CCES_01_04_LysC_Try_1to10_Mixt_2_1raw        1 5327922240
#> 5 121219_S_CCES_01_05_LysC_Try_1to10_Mixt_2_2raw        1 5824830336
#> 6 121219_S_CCES_01_06_LysC_Try_1to10_Mixt_2_3raw        1 5674675584