Protein sample quantification or group quantification

Model-based quantification for each condition or for each biological sample per protein in a targeted Selected Reaction Monitoring (SRM), Data-Dependent Acquisition (DDA or shotgun), and Data-Independent Acquisition (DIA or SWATH-MS) experiment. Quantification takes the processed data set by dataProcess as input and automatically generate the quantification results (data.frame) in a long or matrix format.

quantification(
  data,
  type = "Sample",
  format = "matrix",
  use_log_file = TRUE,
  append = FALSE,
  verbose = TRUE,
  log_file_path = NULL
)

Arguments

data
name of the (processed) data set.

type
choice of quantification. "Sample" or "Group" for protein sample quantification or group quantification.

format
choice of returned format. "long" for long format which has the columns named Protein, Condition, LogIntensities (and BioReplicate if it is subject quantification), NumFeature for number of transitions for a protein, and NumPeaks for number of observed peak intensities for a protein. "matrix" for data matrix format which has the rows for Protein and the columns, which are Groups(or Conditions) for group quantification or the combinations of BioReplicate and Condition (labeled by "BioReplicate"_"Condition") for sample quantification. Default is "matrix"

use_log_file
logical. If TRUE, information about data processing will be saved to a file.

append
logical. If TRUE, information about data processing will be added to an existing log file.

verbose
logical. If TRUE, information about data processing wil be printed to the console.

log_file_path
character. Path to a file to which information about data processing will be saved. If not provided, such a file will be created automatically. If `append = TRUE`, has to be a valid path to a file.

Value

data.frame as described in details.

Details

Sample quantification : individual biological sample quantification for each protein. The label of each biological sample is a combination of the corresponding group and the sample ID. If there are no technical replicates or experimental replicates per sample, sample quantification is the same as run summarization from dataProcess. If there are technical replicates or experimental replicates, sample quantification is median among run quantification corresponding MS runs.

Group quantification : quantification for individual group or individual condition per protein. It is median among sample quantification.

The quantification for endogenous samples is based on run summarization from subplot model, with TMP robust estimation.

Examples

# Consider quantitative data (i.e. QuantData) from a yeast study with ten time points of
# interests, three biological replicates, and no technical replicates which is
# a time-course experiment.
# Sample quantification shows model-based estimation of protein abundance in each biological
# replicate within each time point.
# Group quantification shows model-based estimation of protein abundance in each time point.
QuantData<-dataProcess(SRMRawData, use_log_file = FALSE)

#> INFO  [2021-05-14 22:19:44] ** Features with one or two measurements across runs are removed.
#> INFO  [2021-05-14 22:19:44] ** Fractionation handled.
#> INFO  [2021-05-14 22:19:44] ** Updated quantification data to make balanced design. Missing values are marked by NA
#> INFO  [2021-05-14 22:19:44] ** Log2 intensities under cutoff = 3.776  were considered as censored missing values.
#> INFO  [2021-05-14 22:19:44] ** Log2 intensities = NA were considered as censored missing values.
#> INFO  [2021-05-14 22:19:44] ** Use all features that the dataset originally has.
#> INFO  [2021-05-14 22:19:44] 
#>  # proteins: 2
#>  # peptides per protein: 2-2
#>  # features per peptide: 3-3
#> INFO  [2021-05-14 22:19:44] 
#>                     1 2 3 4 5 6 7 8 9 10
#>              # runs 3 3 3 3 3 3 3 3 3  3
#>     # bioreplicates 3 3 3 3 3 3 3 3 3  3
#>  # tech. replicates 1 1 1 1 1 1 1 1 1  1
#> INFO  [2021-05-14 22:19:44]  == Start the summarization per subplot...
#>   |                                                                              |                                                                      |   0%  |                                                                              |===================================                                   |  50%  |                                                                              |======================================================================| 100%
#> INFO  [2021-05-14 22:19:44]  == Summarization is done.

head(QuantData$FeatureLevelData)

#>   PROTEIN         PEPTIDE TRANSITION               FEATURE LABEL GROUP RUN
#> 1    IDHC ATDVIVPEEGELR_2      y7_NA ATDVIVPEEGELR_2_y7_NA     H     0   1
#> 2    IDHC ATDVIVPEEGELR_2      y7_NA ATDVIVPEEGELR_2_y7_NA     L     1   1
#> 3    IDHC ATDVIVPEEGELR_2      y7_NA ATDVIVPEEGELR_2_y7_NA     H     0   2
#> 4    IDHC ATDVIVPEEGELR_2      y7_NA ATDVIVPEEGELR_2_y7_NA     L     1   2
#> 5    IDHC ATDVIVPEEGELR_2      y7_NA ATDVIVPEEGELR_2_y7_NA     H     0   3
#> 6    IDHC ATDVIVPEEGELR_2      y7_NA ATDVIVPEEGELR_2_y7_NA     L     1   3
#>   SUBJECT FRACTION originalRUN censored  INTENSITY ABUNDANCE newABUNDANCE
#> 1       0        1           1    FALSE 84361.0835 15.855859    15.855859
#> 2       1        1           1    FALSE   215.1353  7.240669     7.240669
#> 3       0        1           2    FALSE 62109.5876 15.801179    15.801179
#> 4       2        1           2    FALSE  1205.2252 10.113738    10.113738
#> 5       0        1           3    FALSE 65114.3646 15.755022    15.755022
#> 6       3        1           3    FALSE  1476.3046 10.292109    10.292109
#>   predicted
#> 1        NA
#> 2        NA
#> 3        NA
#> 4        NA
#> 5        NA
#> 6        NA

# Sample quantification
sampleQuant<-quantification(QuantData, use_log_file = FALSE)
head(sampleQuant)

#>    Protein  1_ReplA   1_ReplB   1_ReplC 10_ReplA 10_ReplB 10_ReplC   2_ReplA
#> 1:    IDHC  5.57705  6.811034  6.909093 12.72029 12.77645 12.71629  6.373069
#> 2:    PMG2 10.75693 10.647513 10.530769 10.40899 10.60595 10.34241 10.057710
#>     2_ReplB   2_ReplC   3_ReplA   3_ReplB   3_ReplC   4_ReplA   4_ReplB
#> 1:  6.59575  6.349843  6.140378  6.600981  7.033944  6.790116  6.174932
#> 2: 10.47072 10.613419 10.667443 10.654398 10.419662 10.773446 10.535321
#>      4_ReplC   5_ReplA   5_ReplB   5_ReplC  6_ReplA   6_ReplB   6_ReplC
#> 1:  7.313115  7.260217  7.209141  6.519457 9.653219  9.542782  9.553527
#> 2: 10.696021 10.461630 10.945504 10.572543 9.943402 10.595059 10.587633
#>     7_ReplA  7_ReplB  7_ReplC  8_ReplA  8_ReplB  8_ReplC  9_ReplA  9_ReplB
#> 1: 12.55638 12.65310 12.55085 12.72204 12.78281 12.80676 12.66386 12.72873
#> 2: 10.56620 10.72296 10.37506 10.22503 10.45177 10.39042 10.27202 10.57575
#>     9_ReplC
#> 1: 12.66681
#> 2: 10.39904

# Group quantification
groupQuant<-quantification(QuantData, type="Group", use_log_file = FALSE)
head(groupQuant)

#>    Protein         1       10         2         3         4         5         6
#> 1:    IDHC  6.811034 12.72029  6.373069  6.600981  6.790116  7.209141  9.553527
#> 2:    PMG2 10.647513 10.40899 10.470722 10.654398 10.696021 10.572543 10.587633
#>           7        8        9
#> 1: 12.55638 12.78281 12.66681
#> 2: 10.56620 10.39042 10.39904