Import DIA-Umpire files

Import DIA-Umpire files

DIAUmpiretoMSstatsFormat(
  raw.frag,
  raw.pep,
  raw.pro,
  annotation,
  useSelectedFrag = TRUE,
  useSelectedPep = TRUE,
  removeFewMeasurements = TRUE,
  removeProtein_with1Feature = FALSE,
  summaryforMultipleRows = max,
  use_log_file = TRUE,
  append = FALSE,
  verbose = TRUE,
  log_file_path = NULL,
  ...
)

Arguments

raw.frag
name of FragSummary_date.xls data, which includes feature-level data.

raw.pep
name of PeptideSummary_date.xls data, which includes selected fragments information.

raw.pro
name of ProteinSummary_date.xls data, which includes selected peptides information.

annotation
name of annotation data which includes Condition, BioReplicate, Run information.

useSelectedFrag
TRUE will use the selected fragment for each peptide. 'Selected_fragments' column is required.

useSelectedPep
TRUE will use the selected peptide for each protein. 'Selected_peptides' column is required.

removeFewMeasurements
TRUE (default) will remove the features that have 1 or 2 measurements across runs.

removeProtein_with1Feature
TRUE will remove the proteins which have only 1 feature, which is the combination of peptide, precursor charge, fragment and charge. FALSE is default.

summaryforMultipleRows
max(default) or sum - when there are multiple measurements for certain feature and certain run, use highest or sum of multiple intensities.

use_log_file
logical. If TRUE, information about data processing will be saved to a file.

append
logical. If TRUE, information about data processing will be added to an existing log file.

verbose
logical. If TRUE, information about data processing wil be printed to the console.

log_file_path
character. Path to a file to which information about data processing will be saved. If not provided, such a file will be created automatically. If `append = TRUE`, has to be a valid path to a file.

...
additional parameters to `data.table::fread`.

Value

data.frame with the required format of MSstats.